Create StarSensor.py
StarSensor.py implements Python code to act as a sensor for the satellite. It initializes with a specified field of view, can be configured via an instance of the Environment class, and services requests to obtain the Magnitude of the brightest star in the field of view, returning it as a 16-bit signed hexidecimal value.