Create Environment.py
Environment.py includes code that implements the Environment class, which is used by StarSensor. Environment keeps track of the simulation clock, which is used to determine the right ascension of the satellite, and the rotational period, which is used to determine the declination of the satellite. Both of these are used by the StarSensor to determine which stars are being observed, which, in turn, allows the system to identify its position.