Create StarCatalog.py
StarCatalog.py includes code that implements the StarCatalog object, which is used by StarSensor. StarCatalog contains a list of Star objects and, when provided a field of view, a right ascension, and a declination, provides the star with the lowest magnitude (i.e. the brightest star) given the arguments. This is used by the StarSensor, in tandem with the right ascension and declination, to identify its position and orientation over the earth.