Create Star.py
Star.py includes code that implements the Star object, which is used by StarCatalog and StarSensor. Star is a structure that contains information on the star's ID number, magnitude, right ascension, and declination. StarCatalog implements Star to generate a list of stars with all the corresponding information. StarSensor, through StarCatalog, identifies the brightest star in the satellite's field of view.